// arg-web-frontend/src/types/accounts-payable.types.ts
export interface ValidationErrors {
  company?: string;
  voucherToPay?: string;
  batchNo?: string;
  bankAccountGL?: string;
  dateToPayBy?: string;
  companyNo?: string;
  year?: string;
}

export interface ReportTableData {
  key: string;
  pdfFileName: string;
  reportDateTime: string;
  reportType?: string;
  reportFileType: string;
  status: string;
  filePath: string;
  formType:string;
  fileName: string;
  fileType: string;
}